# Reduced allocations in StreamCore#step by using subtype polymorphism instead of a catamorphism