--- a/core/shared/src/main/scala/fs2/Stream.scala
+++ b/core/shared/src/main/scala/fs2/Stream.scala
@@ -292,9 +292,9 @@ final class Stream[+F[_],+O] private (private val coreRef: Stream.CoreRef[F,O]) 
 
   def step: Pull[F,Nothing,(NonEmptyChunk[O],Handle[F,O])] =
     Pull.evalScope(get.step).flatMap {
-      case None => Pull.done
-      case Some(Left(err)) => Pull.fail(err)
-      case Some(Right((c, h))) => Pull.pure((c, new Handle(Nil, Stream.mk(h))))
+      case StreamCore.StepResult.Done => Pull.done
+      case StreamCore.StepResult.Failed(err) => Pull.fail(err)
+      case StreamCore.StepResult.Emits(c, h) => Pull.pure((c, new Handle(Nil, Stream.mk(h))))
     }
 
   def stepAsync[F2[_],O2>:O](