There's a small performance improvement as a result but nothing major. There are less allocations so perhaps less GC pressure.

Before:

```
[info] Benchmark                                  Mode  Cnt    Score    Error  Units
[info] StreamBenchmark.awaitPull_256             thrpt   20  170.066 ±  5.208  ops/s
[info] StreamBenchmark.eval_102400               thrpt   20    1.450 ±  0.025  ops/s
[info] StreamBenchmark.leftAssocConcat_102400    thrpt   20    4.836 ±  0.228  ops/s
[info] StreamBenchmark.leftAssocFlatMap_102400   thrpt   20   12.296 ±  0.660  ops/s
[info] StreamBenchmark.rightAssocConcat_102400   thrpt   20    5.095 ±  0.303  ops/s
[info] StreamBenchmark.rightAssocFlatMap_102400  thrpt   20   12.748 ±  0.482  ops/s
[info] StreamBenchmark.runLog_102400             thrpt   20  746.634 ± 22.741  ops/s
```

With this PR:

```
[info] Benchmark                                  Mode  Cnt    Score    Error  Units
[info] StreamBenchmark.awaitPull_256             thrpt   20  182.774 ±  6.579  ops/s
[info] StreamBenchmark.eval_102400               thrpt   20    1.501 ±  0.030  ops/s
[info] StreamBenchmark.leftAssocConcat_102400    thrpt   20    5.044 ±  0.281  ops/s
[info] StreamBenchmark.leftAssocFlatMap_102400   thrpt   20   13.354 ±  0.516  ops/s
[info] StreamBenchmark.rightAssocConcat_102400   thrpt   20    5.171 ±  0.327  ops/s
[info] StreamBenchmark.rightAssocFlatMap_102400  thrpt   20   12.620 ±  0.540  ops/s
[info] StreamBenchmark.runLog_102400             thrpt   20  752.059 ± 20.623  ops/s
```

Review by @pchiusano and @djspiewak