@djspiewak 2016-11-24
Fewer allocations spread across more (nominal) types is *almost* always a win due to the way the JVM is tuned.  Once things have been merged, I'll run @rossabaker's suite and see if we've made measurable progress relative to his benchmarks.